% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomass_mesomicro}
\alias{biomass_mesomicro}
\title{Meso and Micro zooplankton average biomass values}
\format{
a tibble with 44 rows and 4 columns.
\describe{
  \item{Taxname}{Current scientific name.}
  \item{Level}{Taxonomic level of the taxa.}
  \item{Lifestage}{Plankton lifestage.}
  \item{Carbon_mass_micrograms}{Average carbon mass of an individual in micrograms.}
}
}
\usage{
biomass_mesomicro
}
\description{
Average carbon biomass values for meso and micro zooplankton to be used for biomass conversions
}
\seealso{
\code{\link{biomass_macro}}
}
\keyword{datasets}
