% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{completeTaxaList}
\alias{completeTaxaList}
\title{All taxonomic names}
\format{
a character vector of length 454.
}
\usage{
completeTaxaList
}
\description{
A complete list of all valid taxonomic names included in the full dataset. Used to limit choices for filtering by taxa.
}
\seealso{
\code{\link{Taxnamefinder}}, \code{\link{Zoopsynther}}, \code{\link{zooper}}
}
\keyword{datasets}
