% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomic_utilities.R
\name{Datareducer}
\alias{Datareducer}
\title{Find all unique values within a tibble}
\usage{
Datareducer(Data, Reduced_vars)
}
\arguments{
\item{Data}{Data frame containing at least the columns specified in \code{Reduced_vars}.}

\item{Reduced_vars}{Columns from which you would like to select all unique values. Must be quoted names of columns present in \code{Data}.}
}
\value{
A vector of unique non-NA values from the \code{Reduced_vars} columns of \code{Data}.
}
\description{
Outputs a vector with all unique non-NA values in a subset of selected columns from a tibble
}
\examples{
#Find all unique taxonomic names in the crosswalk table.
\dontrun{
All_taxa <- Datareducer(crosswalk, c("Phylum", "Class", "Order", "Family", "Genus", "Species"))
}
}
\seealso{
\code{\link{Zoopsynther}}, \code{\link{crosswalk}}, \code{\link{zoopComb}}
}
\author{
Sam Bashevkin
}
\keyword{internal}
