% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downloader.R
\name{Downloader}
\alias{Downloader}
\title{Download files and retry if failed}
\usage{
Downloader(url, ...)
}
\arguments{
\item{url}{a \code{\link{character}} string (or longer vector e.g.,
    for the \code{"libcurl"} method) naming the URL of a resource to be
    downloaded.}

\item{...}{
  Arguments passed on to \code{\link[utils:download.file]{utils::download.file}}
  \describe{
    \item{\code{destfile}}{a character string (or vector, see \code{url}) with
    the name where the downloaded file is saved.  Tilde-expansion is
    performed.}
    \item{\code{method}}{Method to be used for downloading files.  Current
    download methods are \code{"internal"}, \code{"wininet"} (Windows
    only) \code{"libcurl"}, \code{"wget"} and \code{"curl"}, and there
    is a value \code{"auto"}: see \sQuote{Details} and \sQuote{Note}.

    The method can also be set through the option
    \code{"download.file.method"}: see \code{\link{options}()}.
  }
    \item{\code{quiet}}{If \code{TRUE}, suppress status messages (if any), and
    the progress bar.}
    \item{\code{mode}}{character.  The mode with which to write the file.  Useful
    values are \code{"w"}, \code{"wb"} (binary), \code{"a"} (append) and
    \code{"ab"}.  Not used for methods \code{"wget"} and \code{"curl"}.
    See also \sQuote{Details}, notably about using \code{"wb"} for Windows.
  }
    \item{\code{cacheOK}}{logical.  Is a server-side cached value acceptable?}
    \item{\code{extra}}{character vector of additional command-line arguments for
    the \code{"wget"} and \code{"curl"} methods.}
    \item{\code{headers}}{named character vector of HTTP headers to use in HTTP
    requests.  It is ignored for non-HTTP URLs.  The \code{User-Agent}
    header, coming from the \code{HTTPUserAgent} option (see
    \code{\link{options}}) is used as the first header, automatically.}
  }}
}
\description{
This function will try 3 times to download a file
}
\keyword{internal}
