% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_utilities.R
\name{edi_entity_list}
\alias{edi_entity_list}
\title{Create a list of entities from an EDI package}
\usage{
edi_entity_list(PID)
}
\arguments{
\item{PID}{EDI package ID}
}
\description{
This function lists all data entities for an EDI package
}
\keyword{internal}
