% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_utilities.R
\name{ftp_file_list}
\alias{ftp_file_list}
\title{Create a list of files on an FTP}
\usage{
ftp_file_list(URL)
}
\arguments{
\item{URL}{URL of the FTP site}
}
\description{
This function lists all files from a given FTP URL
}
\keyword{internal}
