% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_utilities.R
\name{html_file_list}
\alias{html_file_list}
\title{Create a list of files from an HTML site}
\usage{
html_file_list(URL)
}
\arguments{
\item{URL}{URL of the site}
}
\description{
This function lists all files from a given URL. It has only been tested for the CDFW filelib https://filelib.wildlife.ca.gov/Public/
}
\keyword{internal}
