% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomic_utilities.R
\name{LCD_Taxa}
\alias{LCD_Taxa}
\title{Apply LCD approach for "Taxa" option}
\usage{
LCD_Taxa(
  df,
  Taxalevel,
  Taxcats_g = c("Genus_g", "Family_g", "Order_g", "Class_g", "Phylum_g")
)
}
\arguments{
\item{df}{Zooplankton dataset including columns named the same as the \code{Taxcats_g}, a \code{Taxname} column, and no other taxonomic identifying columns.}

\item{Taxalevel}{The value of Taxcats_g on which to apply this function.}

\item{Taxcats_g}{The names of taxonomic levels with _g appended corresponding to columns in the zooplankton dataset with higher-level taxonomic categories that have been determined to be important groups for which sums will be calculated. Defaults to \code{Taxcats_g = c("Genus_g", "Family_g", "Order_g", "Class_g", "Phylum_g")}.}
}
\value{
A tibble with sums calculated for each unique value in \code{df$Taxalevel}. Sums will be excluded for grouping taxa that only contain 1 unique Taxname.
}
\description{
Sums to least common denominator taxa, one taxonomic level at a time
}
\details{
This function is designed to work on just one Taxalevel at a time. To apply to multiple Taxalevels, use \link[purrr]{map} or \link[base]{apply} functions to apply across taxonomic levels.
}
\examples{
library(magrittr)
library(rlang)
UniqueTaxa<-zoopComb\%>\%
  dplyr::select(.data$Taxname)\%>\%
  dplyr::distinct()\%>\%
  dplyr::left_join(dplyr::select(crosswalk, .data$Taxname, .data$Level)\%>\%
                     dplyr::distinct(),
                   by="Taxname")\%>\%
  dplyr::filter(.data$Level!="Species")\%>\%
  dplyr::pull(.data$Taxname)
df <- zoopComb\%>\%
  dplyr::mutate_at(c("Genus", "Family", "Order", "Class", "Phylum"),
                   list(g=~dplyr::if_else(.\%in\%UniqueTaxa, ., NA_character_)))\%>\%
  dplyr::select(-.data$Phylum, -.data$Class, -.data$Order, -.data$Family,
                -.data$Genus, -.data$Species, -.data$Taxlifestage)
family_sums <- LCD_Taxa(df, "Family_g")

}
\seealso{
\code{\link{Zoopsynther}}, \code{\link{crosswalk}}, \code{\link{zoopComb}}
}
\author{
Sam Bashevkin
}
\keyword{Taxonomy}
\keyword{zooplankton}
