% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomic_utilities.R
\name{rowAny}
\alias{rowAny}
\title{Find rows with any TRUE values}
\usage{
rowAny(x)
}
\arguments{
\item{x}{Row.}
}
\description{
Helper function to replace \code{dplyr::any_vars}, copied from the \code{dplyr} "colwise" vignette
}
\keyword{internal}
