% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomic_utilities.R
\name{SourceTaxaKeyer}
\alias{SourceTaxaKeyer}
\title{Unique taxa by lifestage combinations present in each source and size class}
\usage{
SourceTaxaKeyer(Data, Crosswalk)
}
\arguments{
\item{Data}{Zooplankton dataset. Must have a column named \code{Source} with the names of the source datasets and a column named \code{SizeClass} with the names of the zooplankton size classes.}

\item{Crosswalk}{Crosswalk table (e.g., \code{\link{crosswalk}}) with columns named "Phylum", "Class", "Order", "Family", "Genus", "Taxname", "Lifestage", and column names corresponding to each unique value of \code{paste(data$Source, data$SizeClass, sep="_")}.}
}
\value{
a tibble with the complete taxonomic information for each combination of source and size class.
}
\description{
Computes a dataframe with all unique taxa by lifestage combinations present in each source and size class
}
\examples{
SourceTaxaKey <- SourceTaxaKeyer(Data = zoopComb, Crosswalk = crosswalk)
}
\seealso{
\code{\link{Zoopsynther}}, \code{\link{crosswalk}}, \code{\link{zoopComb}}
}
\author{
Sam Bashevkin
}
\keyword{Taxonomy,}
\keyword{zooplankton}
