% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{startDates}
\alias{startDates}
\title{Start dates}
\format{
a tibble with 3 columns and 10 rows
\describe{
  \item{Source}{Abbreviated name of the source dataset.  "EMP"=Environmental Monitoring Program, "FRP"=Fish Restoration Program, "FMWT"= Fall Midwater Trawl), "STN"= Townet Survey, and "20mm" =20mm survey.}
  \item{SizeClass}{Net size class. Micro corresponds to 43 \eqn{\mu}m mesh, Meso corresponds to 150-160 \eqn{\mu}m mesh, and Macro corresponds to 500-505 \eqn{\mu}m mesh. However, prior to 1974 EMP macrozooplankton were sampled with a 930 \eqn{\mu}m mesh net.}
  \item{Startdate}{Date first sample was collected.}
}
}
\usage{
startDates
}
\description{
First dates sampled by each survey and size class
}
\seealso{
\code{\link{Uncountedyears}}, \code{\link{Zoopsynther}}
}
\keyword{datasets}
