% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Station locations}
\format{
a tibble with 399 rows and 4 columns
\describe{
  \item{Source}{Abbreviated name of the source dataset}
  \item{Station}{Sampling station name}
  \item{Latitude}{Latitude in decimal degrees}
  \item{Longitude}{Longitude in decimal degrees}
}
}
\usage{
stations
}
\description{
Latitudes and longitudes for each zooplankton station.
}
\seealso{
\code{\link{Zoopdownloader}}, \code{\link{zooper}}
}
\keyword{datasets}
