% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stationsEMPEZ}
\alias{stationsEMPEZ}
\title{EMP EZ Station locations}
\format{
a tibble with 575 rows and 4 columns
\describe{
  \item{Date}{Date sample was collected}
  \item{Station}{Sampling station name}
  \item{Latitude}{Latitude in decimal degrees}
  \item{Longitude}{Longitude in decimal degrees}
}
}
\usage{
stationsEMPEZ
}
\description{
Latitudes and longitudes for EMP EZ stations on each sampling date from 2004 to present.
}
\seealso{
\code{\link{Zoopdownloader}}, \code{\link{zooper}}
}
\keyword{datasets}
