% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_utilities.R
\name{Tryer}
\alias{Tryer}
\title{Try a function call n times before failing}
\usage{
Tryer(n, fun, ...)
}
\arguments{
\item{n}{Number of times to try function before erroring}

\item{fun}{Function to try}

\item{...}{parameters for function}
}
\description{
This function will try n times to run a function call before failing. Mainly used for downloading files and accessing URLs
}
\keyword{internal}
