% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomic_utilities.R
\name{Uncountedyears}
\alias{Uncountedyears}
\title{Detect dates when a species was not counted}
\usage{
Uncountedyears(Source, SizeClass, Crosswalk)
}
\arguments{
\item{Source}{String with the name of the source dataset (e.g., \code{Source="EMP"}).}

\item{SizeClass}{String with the name of the desired zooplankton size class (e.g., \code{Source="Meso"}).}

\item{Crosswalk}{Crosswalk table like \code{\link{crosswalk}} or another table in the same format.}
}
\value{
A tibble with columns for the Taxlifestage, Taxname, Lifestage, Source, Sizeclass, and then a list-column of years in which that particular taxon was not counted in the specified study and size class. Taxa that were counted in all applicable years are not included in the tibble.
}
\description{
Detects years when a species was present in the system (i.e., post-invasion for invasive species) but not counted in each zooplankton survey
}
\details{
This function is designed to work on one source and size class at a time. To apply across multiple IDs, use the \link[purrr]{map} or \link[base]{apply} functions.
}
\examples{
require(purrr)
require(dplyr)

datasets<-zooper::zoopComb\%>\%
 mutate(names=paste(Source, SizeClass, sep="_"))\%>\%
 select(names, Source, SizeClass)\%>\%
 filter(Source\%in\%c("EMP", "FMWT", "twentymm"))\%>\%
 distinct()

BadYears<-map2_dfr(datasets$Source, datasets$SizeClass, Uncountedyears, Crosswalk=zooper::crosswalk)
}
\author{
Sam Bashevkin
}
