% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomic_utilities.R
\name{Wordremover}
\alias{Wordremover}
\title{Remove words from a character vector list element}
\usage{
Wordremover(ID, Taxlifestage_list, Remove_taxa)
}
\arguments{
\item{ID}{The quoted name of an element in both \code{Taxlifestage_list} and \code{Remove_taxa}}

\item{Taxlifestage_list}{The list of character vectors of Taxlifestages from which elements will be removed.}

\item{Remove_taxa}{The list of character vectors of Taxa names containing values to be removed from \code{Taxlifestage_list}.}
}
\value{
A comma-separated string of remaining Taxlifestages from \code{Taxlifestage_list[[ID]]}
}
\description{
Given 2 lists of character vectors and an ID corresponding to the name of an element in each list, this function will, for the named element, remove values in one character vector that correspond to values in the other and output a concatenated string of all remaining values.
}
\details{
This function is designed to work on one ID at a time. To apply across multiple IDs, use the \link[purrr]{map} or \link[base]{apply} functions.
}
\examples{
\dontrun{
Taxlifestage_list <- list(EMP = paste(crosswalk$EMP_Meso[!is.na(crosswalk$EMP_Meso)],
                          crosswalk$Lifestage[!is.na(crosswalk$EMP_Meso)]))
Remove_taxa <- list(EMP = crosswalk$EMP_Micro[!is.na(crosswalk$EMP_Micro)])
Meso_not_Micro <- Wordremover("EMP", Taxlifestage_list, Remove_taxa)
}
}
\seealso{
\code{\link{Zoopsynther}}, \code{\link{crosswalk}}
}
\author{
Sam Bashevkin
}
\keyword{Internal}
