% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{Zoopbiomass}
\alias{Zoopbiomass}
\title{Convert CPUE to biomass
This function converts zooplankton CPUE to carbon biomass (Carbon biomass per unit effort (\eqn{\mu}g/ \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^{3}}})) for taxa with conversion equations.}
\usage{
Zoopbiomass(
  Zoop,
  ZoopLengths = NULL,
  Biomass_mesomicro = zooper::biomass_mesomicro,
  Biomass_macro = zooper::biomass_macro
)
}
\arguments{
\item{Zoop}{Zooplankton count dataset}

\item{ZoopLengths}{Zooplankton length dataset for macrozooplankton. May be left NULL if the Macro size class is not included.}

\item{Biomass_mesomicro}{The micro and meso zooplankton biomass conversion table. The default is \code{\link{biomass_mesomicro}}}

\item{Biomass_macro}{The macro zooplankton biomass conversion table. The default is \code{\link{biomass_macro}}}
}
\description{
Convert CPUE to biomass
This function converts zooplankton CPUE to carbon biomass (Carbon biomass per unit effort (\eqn{\mu}g/ \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^{3}}})) for taxa with conversion equations.
}
