% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooper-package.R
\docType{package}
\name{zooper-package}
\alias{zooper-package}
\title{zooper: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.

This package is used for calculating the number of zooplankton per liter.
The package includes a function used for this calculation and an example
dataset with their respective documentation.
}
\references{
\url{https://github.com/claireestey/zooper}
}
\author{
Claire Estey, Department of Biology, McMaster University \email{esteyc@mcmaster.ca}
}
\keyword{internal}
