% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoopl.R
\name{zoopl}
\alias{zoopl}
\title{Test function}
\usage{
zoopl(a = 1, b = 2)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The product of \code{a} and \code{b}
}
\description{
Test function
}
\examples{
zoopl(2, 5)
zoopl(4, 8)
}
