% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomass_macro}
\alias{biomass_macro}
\title{Macro zooplankton length-weight equations}
\format{
a tibble with 23 rows and 9 columns.
\describe{
  \item{Taxname}{Current scientific name.}
  \item{Level}{Taxonomic level of the taxa.}
  \item{Preservative}{Preservative used to store sample before individuals were meaured to develop the equations.}
  \item{Weight_type}{The type of weight measurement.}
  \item{N}{The number of individuals used in developing the equation.}
  \item{Min_length}{Minimum length (mm) of individuals used in developing the equation.}
  \item{Max_length}{Maximum length (mm) of individuals used in developing the equation.}
  \item{a}{Intercept of the length-weight equation (g).}
  \item{b}{Slope of the length-weight equation (g/mm).}
}
}
\usage{
biomass_macro
}
\description{
Length-weight equations for macro zooplankton to be used for biomass conversions.
The equations relate length in mm to mass in grams.
}
\seealso{
\code{\link{biomass_mesomicro}}
}
\keyword{datasets}
