% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crosswalk}
\alias{crosswalk}
\title{Taxonomic crosswalk among datasets}
\format{
a tibble with 317 rows and 28 variables
\describe{
  \item{EMP_Micro}{Taxonomic codes used in the Environmental Monitoring Program microzooplankton (43 \eqn{\mu}m) mesh dataset}
  \item{EMP_Meso}{Taxonomic codes used in the Environmental Monitoring Program mesozooplankton (160 \eqn{\mu}m) mesh dataset}
  \item{EMP_Macro}{Taxonomic codes used in the Environmental Monitoring Program macrozooplankton (505 \eqn{\mu}m mesh) dataset}
  \item{STN_Meso}{Taxonomic codes used in the Townet Survey mesozooplankton (160 \eqn{\mu}m mesh) dataset}
  \item{STN_Macro}{Taxonomic codes used in the Townet Survey macrozooplankton (505 \eqn{\mu}m mesh) dataset}
  \item{FMWT_Meso}{Taxonomic codes used in the Fall Midwater Trawl mesozooplankton (160 \eqn{\mu}m mesh) dataset}
  \item{FMWT_Macro}{Taxonomic codes used in the Fall Midwater Trawl macrozooplankton (505 \eqn{\mu}m mesh) dataset}
  \item{twentymm_Meso}{Taxonomic codes used in the 20mm Survey mesozooplankton (160 \eqn{\mu}m mesh) dataset}
  \item{FRP_Meso}{Taxonomic codes used in the Fish Restoration Program mesozooplankton (150 \eqn{\mu}m mesh) dataset}
  \item{FRP_Macro}{Taxonomic codes used in the Fish Restoration Program macrozooplankton (500 \eqn{\mu}m mesh) dataset}
  \item{YBFMP}{Taxonomic codes used in the Yolo Bypass Fish Monitoring Program zooplankton dataset)}
  \item{Lifestage}{Plankton lifestage}
  \item{Taxname}{Current scientific name}
  \item{Level}{Taxonomic level of the taxa}
  \item{Phylum}{Phylum}
  \item{Class}{Class}
  \item{Order}{Order}
  \item{Family}{Family}
  \item{Genus}{Genus}
  \item{Species}{Species}
  \item{Intro}{Introduction year for non-native species}
  \item{EMPstart}{First year the Environmental Monitoring Program starting counting this taxa}
  \item{EMPend}{Last year the Environmental Monitoring Program counted this taxa}
  \item{FMWTstart}{First year the Fall Midwater Trawl starting counting this taxa}
  \item{FMWTend}{Last year the Fall Midwater Trawl counted this taxa}
  \item{twentymmstart}{First year the 20mm Survey starting counting this taxa}
  \item{twentymmend}{Last year the 20mm Survey counted this taxa}
  \item{twentymmstart2}{First year the 20mm Survey restarted counting this taxa}
  }
}
\usage{
crosswalk
}
\description{
A crosswalk table relating the taxonomic code used by each dataset to current scientific names, life stages, and taxonomic hierarchies.
}
\seealso{
\code{\link{Zoopdownloader}}, \code{\link{Zoopsynther}}, \code{\link{zooper}}
}
\keyword{datasets}
