% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomic_utilities.R
\name{LCD_Taxa}
\alias{LCD_Taxa}
\title{Apply LCD approach for "Taxa" option}
\usage{
LCD_Taxa(
  Data,
  Taxalevel,
  Groupers = c("Genus_g", "Family_g", "Order_g", "Class_g", "Phylum_g"),
  Response = "CPUE"
)
}
\arguments{
\item{Data}{Zooplankton dataset including columns named the same as the \code{Groupers}, a \code{Taxname} column, "CPUE", and no other taxonomic identifying columns.}

\item{Taxalevel}{The value of Groupers on which to apply this function.}

\item{Groupers}{A character vector of names of additional taxonomic levels to be removed in this step. This vector can include \code{Taxalevel} and, if so, it will be removed from the vector within the function so \code{Taxalevel} is preserved.}

\item{Response}{Which response variable(s) would you like for the zooplankton data? Choices are "CPUE" (catch per unit effort) and "BPUE" (biomass per unit effort). Defaults to \code{Response = "CPUE"}.}
}
\value{
A tibble with sums calculated for each unique value in \code{Data$Taxalevel}. Sums will be excluded for grouping taxa that only contain 1 unique Taxname.
}
\description{
Sums to least common denominator taxa, one taxonomic level at a time
}
\details{
This function is designed to work on just one Taxalevel at a time. To apply to multiple Taxalevels, use \link[purrr]{map} or \link[base]{apply} functions to apply across taxonomic levels.
}
\examples{
\dontrun{
library(dplyr)
df <- zoopComb\%>\%
  mutate(dplyr::across(tidyselect::all_of(c("Genus", "Family", "Order", "Class", "Phylum")),
                   list(g=~if_else(.\%in\%completeTaxaList, ., NA_character_))))\%>\%
  select(-Phylum, -Class, -Order, -Family, -Genus, -Species, -Taxlifestage)
family_sums <- LCD_Taxa(df, "Family_g")
}

}
\seealso{
\code{\link{Zoopsynther}}, \code{\link{crosswalk}}, \code{\link{zoopComb}}
}
\author{
Sam Bashevkin
}
\keyword{Taxonomy}
\keyword{zooplankton}
