% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomic_utilities.R
\name{LCD_Taxa}
\alias{LCD_Taxa}
\title{Apply LCD approach for "Taxa" option}
\usage{
LCD_Taxa(
  df,
  Taxalevel,
  Taxcats_g = c("Genus_g", "Family_g", "Order_g", "Class_g", "Phylum_g")
)
}
\arguments{
\item{df}{Zooplankton dataset including columns named the same as the \code{Taxcats_g}, a \code{Taxname} column, and no other taxonomic identifying columns.}

\item{Taxalevel}{The value of Taxcats_g on which to apply this function.}

\item{Taxcats_g}{The names of taxonomic levels with _g appended corresponding to columns in the zooplankton dataset with higher-level taxonomic categories that have been determined to be important groups for which sums will be calculated. Defaults to \code{Taxcats_g = c("Genus_g", "Family_g", "Order_g", "Class_g", "Phylum_g")}.}
}
\value{
A tibble with sums calculated for each unique value in \code{df$Taxalevel}. Sums will be excluded for grouping taxa that only contain 1 unique Taxname.
}
\description{
Sums to least common denominator taxa, one taxonomic level at a time
}
\details{
This function is designed to work on just one Taxalevel at a time. To apply to multiple Taxalevels, use \link[purrr]{map} or \link[base]{apply} functions to apply across taxonomic levels.
}
\examples{
library(dplyr)
UniqueTaxa<-zoopComb\%>\%
  select(Taxname)\%>\%
  distinct()\%>\%
  left_join(select(crosswalk, Taxname, Level)\%>\%
                     distinct(),
                   by="Taxname")\%>\%
  filter(Level!="Species")\%>\%
  pull(Taxname)
df <- zoopComb\%>\%
  mutate_at(c("Genus", "Family", "Order", "Class", "Phylum"),
                   list(g=~if_else(.\%in\%UniqueTaxa, ., NA_character_)))\%>\%
  select(-Phylum, -Class, -Order, -Family, -Genus, -Species, -Taxlifestage)
family_sums <- LCD_Taxa(df, "Family_g")

}
\seealso{
\code{\link{Zoopsynther}}, \code{\link{crosswalk}}, \code{\link{zoopComb}}
}
\author{
Sam Bashevkin
}
\keyword{Taxonomy}
\keyword{zooplankton}
