% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{startDates}
\alias{startDates}
\title{Start dates}
\format{
a tibble with 3 columns and 10 rows
}
\usage{
startDates
}
\description{
First dates sampled by each survey and size class
}
\seealso{
\code{\link{Uncountedyears}}, \code{\link{Zoopsynther}}
}
\keyword{datasets}
