% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download_utilities.R
\name{zoop_urls}
\alias{zoop_urls}
\title{Extract latest EDI files
This function extracts the latest version of a zooplankton EDI package and the list of files from that package}
\usage{
zoop_urls(Sources)
}
\arguments{
\item{Sources}{Source datasets to be included. Choices include "EMP" (Environmental Monitoring Program), "FRP" (Fish Restoration Program), "FMWT" (Fall Midwater Trawl), "STN" (Townet Survey), "DOP" (Directed Outflow Project), and "20mm" (20mm survey). The YBFMP datasets cannot be used in this function due to taxonomic and life stage issues with that dataset. Defaults to \code{Sources = c("EMP", "FRP", "FMWT", "STN", "20mm", "DOP")}.}
}
\value{
A list with the files and/or URLs for each source dataset
}
\description{
Extract latest EDI files
This function extracts the latest version of a zooplankton EDI package and the list of files from that package
}
\author{
Sam Bashevkin
}
