% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoop_downloader.R
\name{Zoopdownloader}
\alias{Zoopdownloader}
\title{Downloads and combines zooplankton datasets from the Sacramento San Joaquin Delta}
\usage{
Zoopdownloader(
  Data_folder = "data",
  Save_object = TRUE,
  Return_object = FALSE,
  Return_object_type = "List",
  Redownload_data = FALSE,
  Zoop_path = file.path(Data_folder, "zoopforzooper"),
  Env_path = file.path(Data_folder, "zoopenvforzooper")
)
}
\arguments{
\item{Data_folder}{Path to folder in which source datasets are stored, and to which you would like datasets to be downloaded if you set `Redownload_data = TRUE`.}

\item{Save_object}{Should the combined data be saved to disk? Defaults to \code{Save_object = TRUE}.}

\item{Return_object}{Should data be returned as an R object? If \code{TRUE}, the function will return the full combined dataset. Defaults to `Return_object = FALSE`.}

\item{Return_object_type}{If \code{Return_object = TRUE}, should data be returned as a combined dataframe (\code{Return_object_type = "Combined"}) or a list with component "Zooplankton" containing the zooplankton data and component "Environment" containing the environmental data (\code{Return_object_type = "List"}, the default). A list is required to feed data into the \code{zooper} function without saving the combined dataset to disk.}

\item{Redownload_data}{Should source datasets be redownloaded from the internet? Defaults to \code{Redownload_data = FALSE}.}

\item{Zoop_path}{File path specifying the folder and filename of the zooplankton dataset. Defaults to \code{Zoop_path = file.path(Data_folder, "zoopforzooper")}.}

\item{Env_path}{File path specifying the folder and filename of the dataset with accessory environmental parameters. Defaults to \code{Env_path = file.path(Data_folder, "zoopenvforzooper")}.}
}
\value{
If \code{Return_object = TRUE}, returns the combined dataset as a tibble. If \code{Save_object = TRUE}, writes 2 .Rds files to disk: one with the zooplankton catch data and another with accessary environmental parameters.
}
\description{
This function downloads datasets from the internet, converts them to a consistent format, binds them together, and exports the combined dataset as .Rds R data files.
}
\examples{
Data <- Zoopdownloader(Return_object = TRUE, Save_object = FALSE)
}
\author{
Sam Bashevkin
}
\keyword{download,}
\keyword{integration,}
\keyword{synthesis,}
\keyword{zooplankton}
