% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooper-package.R
\docType{package}
\name{zooper-package}
\alias{zooper-package}
\title{zooper: Calculates Zooplankton Community Composition Metrics}
\description{
This package contains functions used for analysing data-sets containing zooplankton community composition information.

This package is used for calculating the number of zooplankton per liter.
The package includes a function used for this calculation and an example
dataset with their respective documentation.
}
\references{
\url{https://github.com/claireestey/zooper}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/claireestey/zooper}
  \item \url{https://claireestey.github.io/zooper/}
  \item Report bugs at \url{https://github.com/claireestey/zooper/issues}
}

}
\author{
\strong{Maintainer}: Claire Estey \email{esteyc@mcmaster.ca} (\href{https://orcid.org/0000}{ORCID})


Claire Estey, Department of Biology, McMaster University \email{esteyc@mcmaster.ca}
}
\keyword{internal}
