% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooper-package.R
\docType{package}
\name{zooper}
\alias{zooper}
\title{zoopl: A package to test if renv is my issue}
\description{
This package is an exercise in checking if renv() is what is causing my
issues with pkgdown for GEOG 712.
}
\references{
\url{https://github.com/claireestey/zooper}
}
\author{
Claire, McMaster University \email{esteyc@mcmaster.ca}
}
