% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoop_synthesizer.R
\name{Zooper}
\alias{Zooper}
\title{Integrates zooplankton datasets in the Sacramento San Joaquin Delta}
\usage{
Zooper(
  Sources = c("EMP", "FRP", "FMWT", "TNS", "20mm"),
  Size_class = c("Micro", "Meso", "Macro"),
  Data = "Community",
  Date_range = c(NA, NA),
  Months = NA,
  Years = NA,
  Sal_bott_range = NA,
  Sal_surf_range = NA,
  Temp_range = NA,
  Lat_range = NA,
  Long_range = NA,
  Reload_data = F,
  Redownload_data = F,
  All_env = T,
  Shiny = F
)
}
\arguments{
\item{Sources}{Source datasets to be included. Choices include "EMP" (Environmental Monitoring Program), "FRP" (Fish Restoration Program), "FMWT" (Fall Midwater Trawl), "TNS" (Townet Survey), and "20mm" (20mm survey). Defaults to \code{Sources = c("EMP", "FRP", "FMWT", "TNS", "20mm")}.}

\item{Size_class}{Zooplankton size classes (as defined by net mesh sizes) to be included in the integrated dataset. Choices include "Micro" (43 \eqn{\mu}m), "Meso" (150 - 160 \eqn{\mu}m), and "Macro" (500-505 \eqn{\mu}m). Defaults to \code{Size_class = c("Micro", "Meso", "Macro")}.}

\item{Data}{What type of data are you looking for? This option allows you to to choose a final output dataset for either community (\code{Data = "Community"}; the default) or Taxa-specific (\code{Data = "Taxa"}) analyses. See below for more explanation.}

\item{Date_range}{Range of dates to include in the final dataset. To filter within a range of dates, include a character vector of 2  dates formatted in the yyyy-mm-dd format exactly, specifying the upper and lower bounds. To specify an infinite upper or lower bound (to include all values above or below a limit) input \code{NA} for that infinite bound. Defaults to \code{Date_range = c(NA, NA)}, which includes all dates.}

\item{Months}{Months to be included in the integrated dataset. If you wish to only include data from a subset of months, input a vector of integers corresponding to the month numbers you wish to be included. Defaults to \code{Months = NA}, which includes all months.}

\item{Years}{Years to be included in the integrated dataset. If you wish to only include data from a subset of years, input a vector of years you wish to be included. Defaults to \code{Years = NA}, which includes all months.}

\item{Sal_bott_range}{Filter the data by bottom salinity values. Include a vector of length 2 specifying the minimum and maximum values you wish to include. To include all values above or below a limit, utilize Inf or -Inf for the upper or lower bound respectively. Defaults to \code{Sal_bott_range = NA}, which includes all bottom salinities.}

\item{Sal_surf_range}{Same as previous, but for surface salinity.}

\item{Temp_range}{Same as \code{Sal_bott_range} but for surface temperature.}

\item{Lat_range}{Latitude range to include in the final dataset. Include a vector of length 2 specifying the minimum and maximum values you wish to include, in decimal degree format. Defaluts to \code{Lat_range = NA}, which includes all latitudes.}

\item{Long_range}{Same as previous, but for longitude. Don't forget that Longitudes should be negative in the Delta!}

\item{Reload_data}{If set to \code{Reload_data = T} runs the \code{Zoopdownloader} function to re-combine source datasets. To include local versions of the datasets without redownloading them from online, set \code{Reload_data = TRUE} and \code{Redownload_data = FALSE}. Defaults to \code{Reload_data= FALSE}}

\item{Redownload_data}{Should data be re-downloaded from the internet? If set to \code{Redownload_data = TRUE}, runs \code{Zoopdownloader(Redownload_data = TRUE)}. Defaults to \code{Redownload_data = FALSE}.}

\item{All_env}{Should all environmental parameters be included? Defaults to \code{All_env = TRUE}.}

\item{Shiny}{Is this function being used within the shiny app? If set to \code{Shiny = TRUE}, outputs a list with the integrated dataset as one component and the caveats as the other component. Defaults to \code{Shiny = FALSE}.}
}
\value{
An integrated zooplankton dataset.
}
\description{
This function returns an integrated zooplankton dataset with taxonomic issues resolved, according to user-specifications, along with important caveats about the data.
}
\details{
This function combines any combination of the zoo datasets (included as parameters) and calculates least common denominator taxa to facilitate comparisons across datasets with differing levels of taxonomic resolution.
}
\section{Data type}{

The \code{Data} parameter toggles between two approaches to resolving differences in taxonomic resolution. If you want all available data on given Taxa, use \code{Data="Taxa"} but if you want to conduct a community analysis, use \code{Data = "Community"}.
Briefly, \code{Data = "Community"} optimizes for community-level analyses by taking all taxa x life stage combinations that are not measured in every input dataset, and summing them up taxonomic levels to the lowest taxonomic level they belong to that is covered by all datasets. Remaining Taxa x life stage combos that are not covered in all datasets up to the phylum level (usually something like Annelida or Nematoda or Insect Pupae) are removed from the final dataset.
\code{Data = "Taxa"} optimizes for the Taxa-level user by maintaining all data at the original taxonomic level (but it outputs warnings for taxa not measured in all datasets, which we call "orphans"). To facilitate comparions across datasets, this option also sums data into general categories that are comparable across all datasets and years: "summed groups." The new variable "Taxatype" identifies which taxa are summed groups (\code{Taxatype = "Summed group"}), which are measured to the species level (\code{Taxatype = "Species"}), and which are higher taxonomic groupings with the species designation unknown: (\code{Taxatype = "UnID species"}).
}

\examples{
MyZoops <- Zooper(Sources = c("EMP", "FRP", "FMWT"), Size_class = "Meso", Date_range=c("1990-10-01", "2000-09-30"))

}
\author{
Sam Bashevkin
}
\keyword{integration,}
\keyword{synthesis,}
\keyword{zooplankton.}
