% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooper.R
\docType{package}
\name{zooper}
\alias{zooper}
\title{zooper: A package for integrating zooplankton datasets from the Sacramento San Joaquin Delta}
\description{
This package contains 2 functions, 3 lookup tables, and 2 built-in pre-combined datasets (one with the zooplankton data and another with the environmental data).
}
\section{zooper functions}{

\itemize{
  \item \code{\link{Zoopsynther}}
  \item \code{\link{Zoopdownloader}}
}
}

\section{zooper lookup tables}{

\itemize{
  \item \code{\link{crosswalk}}
  \item \code{\link{undersampled}}
  \item \code{\link{stations}}
  }
}

\section{zooper pre-combined (with the \code{\link{Zoopdownloader}} function) zooplankton datasets. These may be out of date}{

\itemize{
  \item \code{\link{zoopComb}}
  \item \code{\link{zoopEnvComb}}
}
}

