% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoop_synthesizer.R
\name{Zoopsynther}
\alias{Zoopsynther}
\title{Integrates zooplankton datasets collected by the Interagency Ecological Program from the Sacramento-San Joaquin Delta}
\usage{
Zoopsynther(
  Data_type = NULL,
  Zoop = zooper::zoopComb,
  ZoopEnv = zooper::zoopEnvComb,
  Zoop_path = NULL,
  Env_path = NULL,
  Sources = c("EMP", "FRP", "FMWT", "STN", "20mm", "DOP", "USGS"),
  Size_class = c("Micro", "Meso", "Macro"),
  Time_consistency = FALSE,
  Intro_lag = 2,
  Response = "CPUE",
  Taxa = NULL,
  Date_range = c(NA, NA),
  Months = NA,
  Years = NA,
  Sal_bott_range = NA,
  Sal_surf_range = NA,
  Temp_range = NA,
  Lat_range = NA,
  Long_range = NA,
  Reload_data = F,
  Redownload_data = F,
  All_env = T,
  Shiny = F,
  Crosswalk = zooper::crosswalk,
  Undersampled = zooper::undersampled,
  CompleteTaxaList = zooper::completeTaxaList,
  StartDates = zooper::startDates,
  ...
)
}
\arguments{
\item{Data_type}{What type of data are you looking for? This option allows you to to choose a final output dataset for either community (\code{Data_type = "Community"}; the default) or Taxa-specific (\code{Data_type = "Taxa"}) analyses. NOTE: If you set \code{Data_type="Community"} we do not recommend utilizing the \code{Taxa} argument. See below for more explanation of this argument.}

\item{Zoop}{Zooplankton data. You must provide the "Zooplankton" element from the list returned from \code{\link{Zoopdownloader}(Save_object = FALSE, Return_object = TRUE, Return_object_type="List")}. The default argument provides the built-in (and possibly outdated) version of this combined dataset. If you instead wish to provide paths to saved datasets from the \code{\link{Zoopdownloader}} function, set \code{Data_list = NULL} and provide \code{Zoop_path}.}

\item{ZoopEnv}{Accessory environmental data. You must provide the "Environment" element from the list returned from \code{\link{Zoopdownloader}(Save_object = FALSE, Return_object = TRUE, Return_object_type="List")}. The default argument provides the built-in (and possibly outdated) version of this combined dataset. If you instead wish to provide paths to saved datasets from the \code{\link{Zoopdownloader}} function, set \code{Data_list = NULL} and provide \code{Env_path}.}

\item{Zoop_path}{If you wish to save time by saving the combined zooplankton datasets returned from the \code{zoopdatadownloader} to disk, provider here the path to the combined zooplankton dataset on disk. You must also set \code{Data_list = NULL}.}

\item{Env_path}{If you wish to save time by saving the combined zooplankton datasets returned from the \code{zoopdatadownloader} to disk, provider here the path to the combined accessory environmental data on disk. You must also set \code{Data_list = NULL}.}

\item{Sources}{Source datasets to be included. Choices include "EMP" (Environmental Monitoring Program), "FRP" (Fish Restoration Program), "FMWT" (Fall Midwater Trawl), "STN" (Townet Survey), "DOP" (Directed Outflow Project), USGS north delta zooplankton and flux, and "20mm" (20mm survey). The YBFMP datasets cannot be used in this function due to taxonomic and life stage issues with that dataset. Defaults to \code{Sources = c("EMP", "FRP", "FMWT", "STN", "20mm", "DOP")}.}

\item{Size_class}{Zooplankton size classes (as defined by net mesh sizes) to be included in the integrated dataset. Choices include "Micro" (43 \eqn{\mu}m), "Meso" (150 - 160 \eqn{\mu}m), and "Macro" (500-505 \eqn{\mu}m). Defaults to \code{Size_class = c("Micro", "Meso", "Macro")}.}

\item{Time_consistency}{Would you like to apply a fix to enforce consistent taxonomic resolution over time? Only available for the Community option.}

\item{Intro_lag}{Only applicable if \code{Time_consistency = TRUE}. How many years after a species is introduced should we expect surveys to start counting them? Defaults to 2.}

\item{Response}{Which response variable(s) would you like for the zooplankton data? Choices are "CPUE" (catch per unit effort) and "BPUE" (carbon biomass per unit effort (\eqn{\mu}g/ \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^{3}}})). Defaults to \code{Response = "CPUE"}.}

\item{Taxa}{If you only wish to include a subset of taxa, provide a character vector of the taxa you wish included. This can include taxa of any taxonomic level (e.g., \code{Taxa = "Calanoida"}) to include only calanoids. NOTE: we do not recommend you use this feature AND set \code{Data_type="Community"}. This is better suited to selecting higher-level taxa. If you wish to just include one or a few species, it would be faster to just filter the output of \code{\link{Zoopdownloader}} to include those taxa. Defaults to \code{NULL}, which includes all taxa.}

\item{Date_range}{Range of dates to include in the final dataset. To filter within a range of dates, include a character vector of 2  dates formatted in the yyyy-mm-dd format exactly, specifying the upper and lower bounds. To specify an infinite upper or lower bound (to include all values above or below a limit) input \code{NA} for that infinite bound. Defaults to \code{Date_range = c(NA, NA)}, which includes all dates.}

\item{Months}{Months (as integers) to be included in the integrated dataset. If you wish to only include data from a subset of months, input a vector of integers corresponding to the months you wish to be included. Defaults to \code{Months = NA}, which includes all months.}

\item{Years}{Years to be included in the integrated dataset. If you wish to only include data from a subset of years, input a vector of years you wish to be included. Defaults to \code{Years = NA}, which includes all months.}

\item{Sal_bott_range}{Filter the data by bottom salinity values. Include a vector of length 2 specifying the minimum and maximum values you wish to include. To include all values above or below a limit, utilize Inf or -Inf for the upper or lower bound respectively. Defaults to \code{Sal_bott_range = NA}, which includes all bottom salinities.}

\item{Sal_surf_range}{Same as previous, but for surface salinity.}

\item{Temp_range}{Same as \code{Sal_bott_range} but for surface temperature.}

\item{Lat_range}{Latitude range to include in the final dataset. Include a vector of length 2 specifying the minimum and maximum values you wish to include, in decimal degree format. Defaults to \code{Lat_range = NA}, which includes all latitudes.}

\item{Long_range}{Same as previous, but for longitude. Don't forget that Longitudes should be negative in the Delta!}

\item{Reload_data}{If set to \code{Reload_data = T} runs the \code{\link{Zoopdownloader}} function to re-combine source datasets. To include local versions of the datasets without redownloading them from online, set \code{Reload_data = TRUE} and \code{Redownload_data = FALSE}. Defaults to \code{Reload_data= FALSE}}

\item{Redownload_data}{Should data be re-downloaded from the internet? If set to \code{Redownload_data = TRUE}, runs \code{\link{Zoopdownloader}(Redownload_data=Redownload_data, Zoop_path=Zoop_path, Env_path=Env_path, ...)}. Defaults to \code{Redownload_data = FALSE}.}

\item{All_env}{Should all environmental parameters be included? Defaults to \code{All_env = TRUE}.}

\item{Shiny}{Is this function being used within the shiny app? If set to \code{Shiny = TRUE}, outputs a list with the integrated dataset as one component and the caveats as the other component. Defaults to \code{Shiny = FALSE}.}

\item{Crosswalk}{Crosswalk table to be used for conversions. Must have columns named for each unique combination of source and size class with an underscore separator, as well as all taxonomic levels Phylum through Species, Taxname (full scientific name) and Lifestage. See \code{\link{crosswalk}} (the default) for an example.}

\item{Undersampled}{A table listing the taxonomic names and life stages of plankton undersampled by each net mesh size (i.e. size class). See \code{\link{undersampled}} (the default) for an example.}

\item{CompleteTaxaList}{Character vector of all taxonomic names in source datasets. Defaults to \code{\link{completeTaxaList}}.}

\item{StartDates}{Tibble with the starting dates of each source dataset. Defaults to \code{\link{startDates}}.}

\item{...}{Arguments passed to \code{\link{Zoopdownloader}}.}
}
\value{
An integrated zooplankton dataset.
}
\description{
This function returns an integrated zooplankton dataset with taxonomic issues resolved, according to user-specifications, along with important caveats about the data. It requires the output of the \code{\link{Zoopdownloader}} function to run. This can be provided either as a list or paths to saved .Rds files generated by the \code{\link{Zoopdownloader}} function. The function defaults to loading pre-packaged combined datasets (which may be outdated).
}
\details{
This function combines any combination of the zooplankton datasets (included as parameters)
and calculates least common denominator taxa to facilitate comparisons across datasets with differing
levels of taxonomic resolution. For more information on the source datasets see \code{\link{zooper}}.
}
\section{Data type}{

The \code{Data_type} parameter toggles between two approaches to resolving differences in taxonomic resolution.
If you want all available data on given Taxa, use \code{Data_type="Taxa"} but if you want to conduct a community
analysis, use \code{Data_type = "Community"}.\cr\cr
Briefly, \code{Data_type = "Community"} optimizes for community-level analyses by taking all taxa x life stage
combinations that are not measured in every input dataset, and summing them up taxonomic levels to the lowest
taxonomic level they belong to that is covered by all datasets. Remaining Taxa x life stage combos that are not
covered in all datasets up to the phylum level (usually something like Annelida or Nematoda or Insect Pupae) are
removed from the final dataset. However, some taxa x life stage combos are retained if they are taxonomic levels
higher than species that are counted in some surveys, and a lower taxonomic level within this group is counted in all surveys.
For example, if we had 3 surveys where surveys A and B count \emph{Pseudodiaptomus forbesi}, \emph{Pseudodiaptomus marinus},
and \emph{Pseudodiaptomus} spp. (UnID) but survey C only counts \emph{P. forbesi} and \emph{P. marinus} then the
\emph{Pseudodiaptomus} spp. (UnID) category would be retained after applying the community approach.\cr\cr
\code{Data_type = "Taxa"} optimizes for the Taxa-level user by maintaining all data at the original taxonomic level
(but it outputs warnings for taxa not measured in all datasets, which we call "orphans").
To facilitate comparisons across datasets, this option also sums data into general categories that are comparable
across all datasets and years: "summed groups." The new variable "Taxatype" identifies which taxa are summed groups
(\code{Taxatype = "Summed group"}), which are measured to the species level (\code{Taxatype = "Species"}), and which
are higher taxonomic groupings with the species designation unknown: (\code{Taxatype = "UnID species"}).
}

\examples{
MyZoops <- Zoopsynther(Data_type = "Community",
Sources = c("EMP", "FRP", "FMWT"),
Size_class = "Meso",
Date_range = c("1990-10-01", "2000-09-30"))
}
\seealso{
\code{\link{Zoopdownloader}}, \code{\link{Taxnamefinder}}, \code{\link{SourceTaxaKeyer}}, \code{\link{crosswalk}}, \code{\link{undersampled}}, \code{\link{zoopComb}}, \code{\link{zoopEnvComb}}, \code{\link{zooper}}
}
\author{
Sam Bashevkin
}
\keyword{integration}
\keyword{synthesis}
\keyword{zooplankton.}
